MachineProjects/requirements.txt
jupyter
numpy
pandas
matplotlib
seaborn
scikit-learn
requests
nltk


fastapi
uvicorn
python-dotenv
httpx
pandas
numpy
scipy
scikit-learn

streamlit